requirements.txt
amqp==2.6.1
asgiref==3.5.0
billiard==3.6.4.0
celery==4.4.7
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.2
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
flatbuffers==2.0
gunicorn==20.1.0
invisible-watermark==0.1.5
kombu==4.6.11
numpy==1.22.2
onnx==1.10.2
onnxruntime==1.10.0
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
protobuf==3.19.4
psycopg2==2.9.3
pyparsing==3.0.7
pytz==2021.3
PyWavelets==1.2.0
redis==4.1.4
six==1.16.0
sqlparse==0.4.2
torch==1.10.2
typing_extensions==4.1.1
vine==1.3.0
whitenoise==6.0.0
wrapt==1.13.3
